+++ website/src/content/docs/vscode.mdx
@@ -2,7 +2,7 @@
 title: VSCode Integration
 ---
 
-# {frontMatter.title}
+# {metadata.title}
 
 Use
 [ESLint VSCode extension](https://marketplace.visualstudio.com/items?itemName=dbaeumer.vscode-eslint)
