--- website/src/content/play.mdx
@@ -1,5 +1,0 @@
----
-title: Playground
----
-
-export { PlayPage as default } from '../components/play-page'
